+++ src/components/Header/MobileSearchHeader/MobileSearchHeader.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function MobileSearchHeader() {
+  return <>MobSearc</>;
+}
